Update proj version, make changes to pass stricter ci
