+++ lib/task-spec/src/task-spec/serialization.cc
@@ -1,0 +1,1 @@
+#include "task-spec/serialization.h"
