+++ lib/task-spec/test/CMakeLists.txt
@@ -2,8 +2,8 @@ ff_add_test_executable(
   NAME
     task-spec-tests
   SRC_PATTERNS
-    src/task-spec/dynamic_graph/*.cc
-  PRIVATE_INCLUDE 
+    src/*.cc
+  PRIVATE_INCLUDE
     src/
   DEPS
     doctest
