--- lib/task-spec/test/src/task-spec/op_ordered_slot_signature.cc
@@ -1,10 +1,0 @@
-#include "task-spec/op_ordered_slot_signature.h"
-#include <doctest/doctest.h>
-
-using namespace ::FlexFlow;
-
-TEST_SUITE(FF_TEST_SUITE) {
-  TEST_CASE("get_op_ordered_slot_signature_for_binding") {
-    CHECK_MESSAGE(false, "TODO: get_op_ordered_slot_signature_for_binding");
-  }
-}
