+++ lib/utils/include/utils/full_binary_tree/binary_tree_path_entry.dtg.toml
@@ -10,8 +10,6 @@ features = [
 
 [[values]]
 name = "LEFT_CHILD"
-key = "left"
 
 [[values]]
 name = "RIGHT_CHILD"
-key = "right"
