+++ lib/utils/include/utils/full_binary_tree/full_binary_tree_node_type.dtg.toml
@@ -10,8 +10,6 @@ features = [
 
 [[values]]
 name = "PARENT"
-key = "parent"
 
 [[values]]
 name = "LEAF"
-key = "leaf"
